Moved reducers into src/store/reducers
