+++ src/store/reducers/debtsSlice.ts
@@ -1,5 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { RootState } from '../../app/store';
+import { RootState } from '../store';
 
 export interface Debt {
     id: number;
